Removed average rating field from Professional entity and modified rating field to float datatype in review entity.

+++ backend/src/main/java/io/github/anigaut/adhdresources/review/Review.java
@@ -54,5 +54,5 @@ public class Review {
 
     @NotNull
     @Column(name = "rating")
-    private Short rating;
+    private float rating;
 }
